Rename ActionsMenu component
