+++ src/components/features/VendorsTable.tsx
@@ -1,28 +1,9 @@
-import ActionsMenu from '@/components/features/ActionsMenu'
+import VendorTableActionsMenu from '@/components/features/VendorTableActionsMenu'
 import DataTable from '@/components/shared/DataTable'
-import WarningActionDialog from '../shared/WarningActionDialog'
 import { useTranslation } from 'react-i18next'
-import { useState } from 'react'
 
 export default function VendorsTable() {
   const { t: translate } = useTranslation()
-  const [openDialog, setOpenDialog] = useState(false)
-
-  const handleClickOpen = () => {
-    setOpenDialog(true)
-  }
-
-  const handleClose = () => {
-    setOpenDialog(false)
-  }
-
-  const onDiscardClick = () => {
-    handleClose()
-  }
-
-  const onConfirmClick = () => {
-    handleClose()
-  }
 
   return (
     <DataTable
@@ -41,39 +22,15 @@ export default function VendorsTable() {
           name: 'Bosch',
           vendorNumber: 1,
           markers: 'Масло',
-          actions: (
-            <ActionsMenu
-              actions={[
-                {
-                  label: translate('actionsMenu.options.delete'),
-                  component: (
-                    <WarningActionDialog
-                      open={openDialog}
-                      title={translate('deleteAction.title')}
-                      content={translate('deleteAction.message')}
-                      discardText={translate('deleteAction.labels.discard')}
-                      confirmText={translate('deleteAction.labels.confirm')}
-                      onCloseDialog={handleClose}
-                      onDiscardClick={onDiscardClick}
-                      onConfirmClick={onConfirmClick}
-                    />
-                  )
-                }
-              ]}
-              handleClickOpen={handleClickOpen}
-            />
-          )
+          actions: <VendorTableActionsMenu />
+        },
+        {
+          name: 'Valeo',
+          vendorNumber: 2,
+          markers: 'Чистачки',
+          actions: <VendorTableActionsMenu />
         }
-        // ,
-        // {
-        //   name: 'Valeo',
-        //   vendorNumber: 2,
-        //   markers: 'Чистачки',
-        //   actions: <ActionsMenu actions={[{ label: translate('actionsMenu.options.delete') }]} />
-        // }
       ]}
     />
   )
 }
-
-// const options = [translate('actionsMenu.options.edit'), translate('actionsMenu.options.delete')]
